# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost","description":"Generated server url"}],"paths":{"/api/v1/verification/vehicle-rc":{"post":{"tags":["verification-controller"],"operationId":"verifyVehicleRC","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RCRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/v1/verification/driving-license":{"post":{"tags":["verification-controller"],"operationId":"verifyDrivingLicense","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DLRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/v1/verification/brands/webhooks/penny-drop":{"post":{"tags":["brand-verification-controller"],"operationId":"handlePennyDropWebhook","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PennyDropWebhook"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/verification/brands/gstin":{"post":{"tags":["brand-verification-controller"],"operationId":"verifyGstin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GstinRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/verification/brands/bank-account":{"post":{"tags":["brand-verification-controller"],"operationId":"verifyBankAccount","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BankAccountRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/verification/biometric":{"post":{"tags":["verification-controller"],"operationId":"verifyBiometric","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BiometricRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/v1/verification/bank-account":{"post":{"tags":["verification-controller"],"operationId":"verifyBankAccount_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BankRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/v1/verification/upload-url":{"get":{"tags":["verification-controller"],"operationId":"getPresignedUploadUrl","parameters":[{"name":"docType","in":"query","required":true,"schema":{"type":"string","enum":["AADHAAR","PAN","DRIVING_LICENSE","RC","SELFIE","GSTIN","FSSAI"]}},{"name":"contentType","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/v1/verification/status/{executiveId}":{"get":{"tags":["verification-controller"],"operationId":"getVerificationSummary","parameters":[{"name":"executiveId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/v1/verification/download-url":{"get":{"tags":["verification-controller"],"operationId":"getPresignedDownloadUrl","parameters":[{"name":"objectKey","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"RCRequest":{"required":["registrationNumber"],"type":"object","properties":{"registrationNumber":{"type":"string"},"documentUrl":{"type":"string"}}},"DLRequest":{"required":["dateOfBirth","dlNumber"],"type":"object","properties":{"dlNumber":{"type":"string"},"dateOfBirth":{"type":"string"},"documentUrl":{"type":"string"}}},"PennyDropWebhook":{"type":"object","properties":{"brandId":{"type":"string","format":"uuid"},"beneficiaryName":{"type":"string"},"status":{"type":"string"},"registeredBrandName":{"type":"string"}}},"GstinRequest":{"required":["brandName","gstin"],"type":"object","properties":{"brandId":{"type":"string","format":"uuid"},"gstin":{"type":"string"},"brandName":{"type":"string"}}},"BankAccountRequest":{"required":["accountNumber","brandName","ifscCode"],"type":"object","properties":{"brandId":{"type":"string","format":"uuid"},"accountNumber":{"type":"string"},"ifscCode":{"type":"string"},"brandName":{"type":"string"}}},"BiometricRequest":{"required":["selfieUrl"],"type":"object","properties":{"selfieUrl":{"type":"string"}}},"BankRequest":{"required":["accountNumber","ifscCode","kycFullName"],"type":"object","properties":{"accountNumber":{"type":"string"},"ifscCode":{"type":"string"},"kycFullName":{"type":"string"}}}}}}